+++ tests/loom.sh
@@ -11,8 +11,8 @@
 # LOOM_LOG=debug
 # LOOM_CHECKPOINT_FILE=target/loom-checkpoint.json
 
-time RUSTFLAGS="${RUSTFLAGS} --cfg loom -C debug-assertions" \
+time RUSTFLAGS="${RUSTFLAGS} --cfg idr_ebr_loom -C debug-assertions" \
      LOOM_MAX_PREEMPTIONS="${LOOM_MAX_PREEMPTIONS:-2}" \
      LOOM_CHECKPOINT_INTERVAL="${LOOM_CHECKPOINT_INTERVAL:-1}" \
      LOOM_LOCATION=1 \
-     cargo test --release --test loom "$@"
+     cargo test --release --features loom --test loom "$@"